# service/docs/swagger.yaml
openapi: 3.0.3
info:
  title: The Zula App
  description: |-
     Organize your thoughts by writing notes, uploading files, saving bookmarks and searching easily on any device with a browser. 
  contact:
    email: alperkilci@gmail.com
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  version: 1.0.0
  
paths:
  /api/v1/ws/{user}:
    get:
      operationId: connectWs
      summary: Connect to websocket
      parameters:
        - in: path
          name: user
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
  /api/v1/users/{shortId}:
    get:
      operationId: getUser
      summary: Get a user by ID
      parameters:
        - in: path
          name: shortId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
    put:
      operationId: updateUser
      summary: Update a user by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userInput'
      parameters:
        - in: path
          name: shortId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user' 

  /api/v1/users/{shortId}/activity:
    get:
      operationId: getUserActivity
      summary: Get user activity
      parameters:
        - in: path
          name: shortId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: sortBy
          schema:
            type: string
        - in: query
          name: sortDirection
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userActivityResult'

  /api/v1/users/{shortId}/insights:
    get:
      operationId: getInsights
      summary: Get dashboard insights
      parameters:
        - in: path
          name: shortId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/insights'

  /api/v1/tags:
    get:
      operationId: getTags
      summary: Get tags
      parameters:
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/tag'

  /api/v1/notes:
    get:
      operationId: getNotes
      summary: List notes
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: sortBy
          schema:
            type: string
        - in: query
          name: sortDirection
          schema:
            type: string
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/noteSearchResult'
    post:
      operationId: createNote
      summary: Create a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/noteInput'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/note'

  /api/v1/notes/{shortId}:
    get:
      operationId: getNote
      summary: Get a note by shortId
      parameters:
        - in: path
          name: shortId
          required: true
          schema:
            type: string
        - in: query
          name: loadDraft
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/note'
    put:
      operationId: updateNote
      summary: Update a note by shortId
      parameters:
        - in: path
          name: shortId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/noteInput'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: boolean
    delete:
      operationId: deleteNote
      summary: Delete a note by shortId
      parameters:
        - in: path
          name: shortId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: boolean

  /api/v1/notes/{shortId}/draft:
    put:
      operationId: saveNoteDraft
      summary: Save draft of a note by notes shortId
      parameters:
        - in: path
          name: shortId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/noteInput'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: boolean
    delete:
      operationId: deleteNoteDraft
      summary: Delete a notes draft by note shortId
      parameters:
        - in: path
          name: shortId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: boolean                

  /api/v1/bookmarks:
    get:
      operationId: getBookmarks
      summary: List bookmarks
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: sortBy
          schema:
            type: string
        - in: query
          name: sortDirection
          schema:
            type: string
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bookmarkSearchResult'
    post:
      operationId: createBookmark
      summary: Create a new bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bookmarkInput'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bookmark'

  /api/v1/bookmarks/{shortId}:
    get:
      operationId: getBookmark
      summary: Get a bookmark by shortId
      parameters:
        - in: path
          name: shortId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bookmark'
    put:
      operationId: updateBookmark
      summary: Update a bookmark by shortId
      parameters:
        - in: path
          name: shortId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bookmarkInput'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: boolean
    delete:
      operationId: deleteBookmark
      summary: Delete a bookmark by shortId
      parameters:
        - in: path
          name: shortId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: boolean

components:
  schemas:
    tag:
      type: object
      properties:
        typeOfParent:
          type: string
        value:
          type: string
        frequency:
          type: integer

    paginationMeta:
      type: object
      properties:
        range:
          type: string
        count:
          type: integer
        hasNextPage:
          type: boolean
        page:
          type: integer
        pageSize:
          type: integer
        sortBy:
          type: string
        sortDirection:
          type: string

    noteInput:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string
        title:
          type: string
        content:
          type: string

    bookmarkInput:
      type: object
      required:
        - url
      properties:
        url:
          type: string
        title:
          type: string
        tags:
          type: array
          items:
            type: string

    userInput:
      type: object
      properties:
        fullname:
          type: string
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string
        language:
          type: string
        theme:
          type: string

    noteLite:
      type: object
      properties:
        shortId:
          type: string
        title:
          type: string

    noteReferenceLink:
      type: object
      properties:
        source:
          type: string
        target:
          type: string

    noteReferences:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/paginationMeta'
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/noteLite'
        links:
          type: array
          items:
            $ref: '#/components/schemas/noteReferenceLink'

    note:
      type: object
      properties:
        shortId:
          type: string
        tags:
          type: array
          items:
            type: string
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
        createdBy:
          type: string
        updatedAt:
          type: string
        updatedBy:
          type: string
        hasDraft:
          type: boolean
        references:
          $ref: '#/components/schemas/noteReferences'

    pageContent:
      type: object
      properties:
        url:
          type: string
        title:
          type: string
        author:
          type: string
        length:
          type: integer
        siteName:
          type: string
        image:
          type: string
        favicon:
          type: string
        mdContent:
          type: string

    bookmark:
      type: object
      required:
        - shortId
        - createdAt
        - updatedAt
        - url
      properties:
        shortId:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
        url:
          type: string
        title:
          type: string
        pageContent:
          $ref: '#/components/schemas/pageContent'
        faviconUrl:
          type: string
        tags:
          type: array
          items:
            type: string

    user:
      type: object
      properties:
        shortId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
        email:
          type: string
        createdAt:
          type: string
        language:
          type: string
        theme:
          type: string

    userActivity:
      type: object
      properties:
        clientId:
          type: string
        resourceType:
          type: string
        action:
          type: string
        objectId:
          type: string
        timestamp:
          type: string

    noteSearchResult:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/paginationMeta'
        items:
          type: array
          items:
            $ref: '#/components/schemas/note'

    bookmarkSearchResult:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/paginationMeta'
        items:
          type: array
          items:
            $ref: '#/components/schemas/bookmark'

    userActivityResult:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/paginationMeta'
        items:
          type: array
          items:
            $ref: '#/components/schemas/userActivity'

    visitingStatistics:
      type: object
      properties:
        typename:
          type: string
        id:
          type: string
        title:
          type: string
        name:
          type: string

    mostVisited:
      allOf:
        - $ref: '#/components/schemas/visitingStatistics'
        - type: object
          properties:
            count:
              type: number

    activityOnDate:
      type: object
      properties:
        date:
          type: string
          format: date
        count:
          type: number

    insights:
      type: object
      properties:
        numberOfNotes:
          type: number
        numberOfBookmarks:
          type: number
        lastVisited:
          type: array
          items:
            $ref: '#/components/schemas/visitingStatistics'
        mostVisited:
          type: array
          items:
            $ref: '#/components/schemas/mostVisited'
        activityGraph:
          type: array
          items:
            $ref: '#/components/schemas/activityOnDate'